requirements.txt
anyio==3.6.2
arabic-reshaper==3.0.0
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect2==0.7.0
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
oscrypto==1.3.0
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.7
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.29.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.2
Whoosh==2.7.4
xhtml2pdf==0.2.10
